requirements.txt
absl-py==2.3.1
accelerate==1.10.1
acvl_utils==0.2.5
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==0.5.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
apache-tvm-ffi==0.1.2
astor==0.8.1
attrs==25.4.0
av==16.0.1
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
blake3==1.0.8
blosc2==3.10.0
brotli==1.2.0
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.1
cloudpickle==3.1.1
compressed-tensors==0.12.2
connected-components-3d==3.25.0
contourpy==1.3.3
cuda-bindings==13.0.3
cuda-pathfinder==1.3.2
cuda-python==13.0.3
cupy-cuda12x==13.6.0
cycler==0.12.1
depyf==0.20.0
diffusers==0.36.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dynamic_network_architectures==0.4.2
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.4.0
fastar==0.6.0
fastrlock==0.8.3
ffmpy==1.0.0
fft-conv-pytorch==1.2.0
filelock==3.19.1
flashinfer-python==0.5.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
gguf==0.17.1
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.75.1
h11==0.16.0
h5py==3.14.0
hdf5storage==0.2.2
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.1
interegular==0.3.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.1
kornia_rs==0.1.9
lark==1.2.2
lazy_loader==0.4
lightning-utilities==0.15.2
linecache2==1.0.0
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
mistral_common==1.8.5
ml_collections==1.1.0
model-hosting-container-standards==0.1.8
modelscope==1.32.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
MultiScaleDeformableAttention==1.0
natsort==8.4.0
ndindex==1.10.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nltk==3.2.4
numba==0.61.2
numexpr==2.13.1
numpy==2.1.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.16.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.2.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.8.1
openai-harmony==0.0.8
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.4
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post7
peft==0.18.1
pillow==11.3.0
pip==25.2
platformdirs==4.5.0
preprocessing==0.1.13
prometheus_client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
pudb==2025.1.5
py-cpuinfo==9.0.0
pybase64==1.4.2
pycountry==24.6.1
pydantic==2.12.4
pydantic_core==2.41.5
pydantic-extra-types==2.10.6
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytorch-lightning==1.9.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
ray==2.51.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.29.0
ruff==0.14.5
safehttpx==0.1.7
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.45.0
setproctitle==1.3.7
setuptools==78.1.1
sewar==0.4.6
shapely==2.1.2
shellingham==1.5.4
simpleitk==2.5.2
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
sphinx-rtd-theme==0.2.4
starlette==0.49.3
stringzilla==4.6.0
submitit==1.5.3
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.9.30
tiktoken==0.12.0
timm==1.0.20
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.0
torchaudio==2.9.0
torchmetrics==1.8.2
torchvision==0.24.0
tqdm==4.67.1
traceback2==1.4.0
transformers==4.57.0
triton==3.5.0
typer==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
unittest2==1.1.0
urllib3==2.5.0
urwid==3.0.5
urwid_readline==0.15.1
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.1
watchfiles==1.1.1
wcwidth==0.5.3
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xformers==0.0.33.post1
xgrammar==0.1.25
yarl==1.22.0
zipp==3.23.0
